Add __init__.py, __version__.py Rename module to src
