+++ refal/__version__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+__version__ = ('1', '1', '0b1')
